requirements.txt
mpmath==1.3.0
pycryptodome==3.23.0
sympy==1.14.0